backend/requirements.txt
fastapi==0.111.0
uvicorn==0.25.0
python-dotenv==1.0.0
supabase==2.5.1
openai==2.3.0